Split gemmi wrappers and fix unbound types for --emit-tsd
Split monolithic gemmi-wrappers.cc (~3000 lines) into:
- gemmi-wrappers-helpers.h: shared includes, structs, helper functions
- gemmi-wrappers.cc: core bindings (Structure, Model, geometry, symmetry, metadata, restraints)
- gemmi-cif-wrappers.cc: CIF doc/block/table, ChemComp, MonLib, Selection
- gemmi-data-wrappers.cc: MTZ, ReflnBlock, Grid, Ccp4, SmallStructure, value_arrays

Fix unbound types blocking --emit-tsd:
- Add value_array for std::array<int,3> (Miller), <double,3>, <double,2>,
  <int,5>, <size_t,3>, <string,2>, and nested array types for Op::Rot/Seitz
- Add register_vector for FTransform, ChemComp::Atom, int8_t, Miller
- Fix static methods incorrectly bound with .function() instead of
  .class_function(): Entity::first_mon, Restraints::lexicographic_str,
  Op::divide_hkl_by_DEN, GroupOps::has_phase_shift, ChemComp::read_group,
  ChemComp::is_peptide_group/is_nucleotide_group, MonLib::relative_monomer_path
- Add register_vector<coot::geometry_distortion_info_t> in moorhen-wrappers.cc

All 16 existing gemmi tests pass. More register_vector fixes needed in
moorhen-wrappers.cc before --emit-tsd will fully succeed.
